+++ pybamm/input/parameters/lithium_ion/NCA_Kim2011.py
@@ -379,8 +379,6 @@ def get_parameter_values():
         "Negative particle radius [m]": 5.083e-07,
         "Negative electrode Bruggeman coefficient (electrolyte)": 2.0,
         "Negative electrode Bruggeman coefficient (electrode)": 2.0,
-        "Negative electrode cation signed stoichiometry": -1.0,
-        "Negative electrode electrons in reaction": 1.0,
         "Negative electrode charge transfer coefficient": 0.5,
         "Negative electrode double-layer capacity [F.m-2]": 0.2,
         "Negative electrode exchange-current density [A.m-2]"
@@ -399,8 +397,6 @@ def get_parameter_values():
         "Positive particle radius [m]": 1.633e-06,
         "Positive electrode Bruggeman coefficient (electrolyte)": 2.0,
         "Positive electrode Bruggeman coefficient (electrode)": 2.0,
-        "Positive electrode cation signed stoichiometry": -1.0,
-        "Positive electrode electrons in reaction": 1.0,
         "Positive electrode charge transfer coefficient": 0.5,
         "Positive electrode double-layer capacity [F.m-2]": 0.2,
         "Positive electrode exchange-current density [A.m-2]"