+++ pybamm/input/parameters/lithium_ion/Ramadass2004.py
@@ -449,8 +449,6 @@ def get_parameter_values():
         "Negative particle radius [m]": 2e-06,
         "Negative electrode Bruggeman coefficient (electrolyte)": 4.0,
         "Negative electrode Bruggeman coefficient (electrode)": 4.0,
-        "Negative electrode cation signed stoichiometry": -1.0,
-        "Negative electrode electrons in reaction": 1.0,
         "Negative electrode charge transfer coefficient": 0.5,
         "Negative electrode double-layer capacity [F.m-2]": 0.2,
         "Negative electrode exchange-current density [A.m-2]"
@@ -470,8 +468,6 @@ def get_parameter_values():
         "Positive particle radius [m]": 2e-06,
         "Positive electrode Bruggeman coefficient (electrolyte)": 4.0,
         "Positive electrode Bruggeman coefficient (electrode)": 4.0,
-        "Positive electrode cation signed stoichiometry": -1.0,
-        "Positive electrode electrons in reaction": 1.0,
         "Positive electrode charge transfer coefficient": 0.5,
         "Positive electrode double-layer capacity [F.m-2]": 0.2,
         "Positive electrode exchange-current density [A.m-2]"